qemu: Fix performance downgrade when running multiple VM instances
This PR is used to fix the issue of #2636.

On AArch64 platform, it doesn't have an architected methods to obtain
entropy, e.g, via instructions, nor does it have real hardware rng
device, so we don't need to provide a "virtio-rng-pci" as a paravirtualized
device to expose entropy inside the virtual machine.

Signed-off-by: Dennis Chen <dennis.chen@arm.com>