After partitioning we now retry the stat prior to formatting
When waiting for devices to refresh a stat can fail with the
error message 'stat: /dev/vda1: stat: no such file or directory'.

This was observered with empty raw disks (*.img) that was being
created by the moby/hyperkit go library.

This commit restores the previous logic of refreshing with mdev
instead of returning an error.

Signed-off-by: dave protasowski <dprotaso@gmail.com>