cmd/packet: Use IPXEScriptURL instead of passing script via Userdata
Using the userdat to pass the iPXE script precludes us from passing
actual userdata to the instance. Instead pass a URL to the iPXE script.
The script is expected to be a <base-url>/<name>-packet.ipxe.

When starting a server also serve the iPXE script from memory (not
the file system). The http server needed extending to differentiate
between the serving files and the iPXE script and ServeMux is used
to differentiate based on path.

A subsequent commit will provide an option to write the iPXE script
to disk.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>