tools/alpine: Build our own wireguard-tools package
Currently, there is a conflict in musl versions between stable
(used for tools/alpine) and edge (where wireguard-tools is).
This cased the tools/alpine build to fail.

With this commit we build our own wireguard-tools package,
using the  APKBUILD file from edge, against the libraries
libraries from stable. We then add the wireguard-tools package
to the mirror.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>