Convert slow net subsystems to modules on 4.14.
In setup_net() there are a few particularly slow subsystems that
contribute more than 140ms of time to the new net namespace creation
path.  The docker daemon doesn't depend on these, and won't modprobe
them into the kernel.  Convert these to modules to reduce the amount of
time it takes for docker to start a container.  This change takes an
additional ~120 ms of time off container start time.

Signed-off-by: Krister Johansen <krister.johansen@oracle.com>