test: Add GCP Run Test
This commit adds the GCP test that formerly ran in LinuxKitCI to run
under rtf.

As GitHub Actions doesn't currently support adding secret files, I've
skipped this test for now. Credentials can be passed via environment
variable but as RTF runs with `-x` the contents is viewable in the logs.

I will create an issue to follow up and find either a way of writing the
variable to file that doesn't compromise it. Or perhaps another approach
that is more compatible with GH actions

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>