Added prefix functionality
The initrd image can now be generated with a prefix path to the root
filesystem of the Docker image, using the '-prefix' option.

This is a useful feature when trying to create a custom ramfs build, as
it removes the constraint of placing the Docker filesystem at the root
of the resulting image.

Signed-off-by: Bogdan Neacsu <bnneacsu@amazon.com>