tests: Fix kernel builder tests
Some kernels are only build for some architectures. The
test assumed that all kernels were build for all architectures.

Now, get a list of architectures for which we have a given
kernel image and then make sure the builder images pointed
to by the label and the builder image tagged by convention
exist and point to the same thing.

Signed-off-by: Rolf Neugebauer <rn@rneugeba.io>