Get BCC building again.
This moves up to bcc 0.20.0 and builds on the latest 3.13 Alpine base
image.  It uses libelf from Alpine, which allows us to drop a number of
the patches we were carrying and reduce the number of steps taken in the
bcc build.

This builds for me on a branch of tip against 5.11.x, 5.10.x,
5.10.x-dbg, and 5.4.x on x86_65.  I have not had a chance to attempt
this on other platforms due to lack of hardware.

Signed-off-by: Krister Johansen <krister.johansen@oracle.com>