pkglib: put real pkg-build-arg-hash into pkg hash
when using
```
buildArgs:
  - REL_HASH_%=@lkt:pkgs:../*
```
then exactly this string is put into the pkg hash, but
if one of these packages are updated, the pkg hash is
not updated, as the string is static.

Instead put in the real pkg-build-arg-hash so that even if
a pkg is updated, it's descendants end up with a new hash.

Signed-off-by: Christoph Ostarek <christoph@zededa.com>